# docs/architecture-pack/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Farmmap API
  version: 1.0.0
  description: >
    Farmmap public and authenticated API covering all Must-Have features (F1–F16).
    All endpoints require a Supabase JWT Bearer token where marked. directory_id
    is resolved server-side from the Host header; it is never a client parameter
    on public endpoints.
  contact:
    name: Farmmap Engineering
    url: https://farmmap.co.uk
servers:
  - url: https://farmmap.co.uk/api
    description: Production
  - url: https://staging.farmmap.co.uk/api
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from auth.getSession()
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Listing not found
        details:
          type: object
          additionalProperties: true
...
    OrderStatus:
      type: string
      enum: [pending, accepted, preparing, dispatched, delivered, cancelled, refunded]
    OrderItem:
      type: object
      required: [product_id, product_name, quantity, unit_price_pence, line_total_pence]
      properties:
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_pence:
          type: integer
        line_total_pence:
          type: integer
    Order:
      type: object
      required: [id, status, subtotal_pence, total_pence, items, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotal_pence:
          type: integer
        commission_pence:
          type: integer
        delivery_fee_pence:
          type: integer
        total_pence:
          type: integer
        currency:
          type: string
          enum: [GBP, EUR]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        delivery_slot:
          type: object
          nullable: true
          properties:
            date:
              type: string
              format: date
            time_from:
              type: string
            time_to:
              type: string
        notes:
          type: string
          nullable: true
        auto_cancel_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      summary: Create an order (consumer checkout)
      description: >
        Creates an order and reserves stock for 15 minutes. Returns a Stripe
        Payment Intent client secret for the consumer to complete payment via
        Stripe Elements. Idempotency key required in request body.
        The listing must be Silver or Gold tier.
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id, items, delivery_zone_id, delivery_slot, buyer_name, buyer_email, idempotency_key]
              properties:
                listing_id:
                  type: string
                  format: uuid
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [product_id, quantity]
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                delivery_zone_id:
                  type: string
                  format: uuid
                delivery_slot:
                  type: object
                  required: [date, time_from, time_to]
                  properties:
                    date:
                      type: string
                      format: date
                    time_from:
                      type: string
                    time_to:
                      type: string
                buyer_name:
                  type: string
                  maxLength: 255
                buyer_email:
                  type: string
                  format: email
                buyer_phone:
                  type: string
                delivery_address:
                  type: object
                  required: [line1, town, postcode]
                  properties:
                    line1:
                      type: string
                    line2:
                      type: string
                    town:
                      type: string
                    county:
                      type: string
                    postcode:
                      type: string
                notes:
                  type: string
                  maxLength: 500
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Order created; payment intent ready
          content:
            application/json:
              schema:
                type: object
                required: [order_id, payment_intent_client_secret, total_pence]
                properties:
                  order_id:
                    type: string
                    format: uuid
                  payment_intent_client_secret:
                    type: string
                    description: Stripe PaymentIntent client_secret for Stripe Elements
                  total_pence:
                    type: integer
                  currency:
                    type: string
                    enum: [GBP, EUR]
        '400':
          description: Delivery postcode not in zone, stock unavailable, or listing not Silver/Gold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate idempotency key — returns existing order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...